+++ src/mons/ghouliath/WitherAway.sol
@@ -39,7 +39,7 @@ contract WitherAway is StandardAttack {
         uint256 attackerPlayerIndex,
         uint256 attackerMonIndex,
         uint256 defenderMonIndex,
-        uint240 extraData,
+        uint16 extraData,
         uint256 rng
     ) public override {
         // Deal the damage and inflict panic
