+++ src/mons/xmon/NightTerrors.sol
@@ -36,7 +36,7 @@ contract NightTerrors is IMoveSet, BasicEffect {
         uint256 attackerPlayerIndex,
         uint256 attackerMonIndex,
         uint256,
-        uint240,
+        uint16,
         uint256
     ) external {
         uint256 defenderPlayerIndex = (attackerPlayerIndex + 1) % 2;
@@ -96,7 +96,7 @@ contract NightTerrors is IMoveSet, BasicEffect {
         return MoveClass.Special;
     }
 
-    function isValidTarget(IEngine, bytes32, uint240) external pure returns (bool) {
+    function isValidTarget(IEngine, bytes32, uint16) external pure returns (bool) {
         return true;
     }
