+++ test/mocks/InvalidMove.sol
@@ -15,7 +15,7 @@ contract InvalidMove is IMoveSet {
         return "Effect Attack";
     }
 
-    function move(IEngine, bytes32, uint256, uint256, uint256, uint240, uint256) external pure {
+    function move(IEngine, bytes32, uint256, uint256, uint256, uint16, uint256) external pure {
         // No-op
     }
 
@@ -31,7 +31,7 @@ contract InvalidMove is IMoveSet {
         return Type.Fire;
     }
 
-    function isValidTarget(IEngine, bytes32, uint240) external pure returns (bool) {
+    function isValidTarget(IEngine, bytes32, uint16) external pure returns (bool) {
         return false;
     }
